pdsensorvis/pyproject.toml
[tool.poetry]
name = "pdsensorvis"
version = "0.1.0"
description = "This is my master's project for analyzing Parkinson's Disease symptoms in a residential environment."
authors = ["Mickey Kim"]
license = "MIT"

[tool.poetry.dependencies]
python = "^3.7"
Django = "^3.2.5"
django-modeladmin-reorder = "^0.3.1"
djangorestframework = "^3.12.4"
django-chartjs = "^2.2.1"
xlwt = "^1.3.0"
whitenoise = "^5.3.0"

[tool.poetry.dev-dependencies]
black = "^22.1.0"
pylint = "^2.12.2"
flake8 = "^4.0.1"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
